more atomic state structure
